+++ cucumber-test/src/main/java/decks/SubDeck.java
@@ -1,7 +1,8 @@
 package decks;
+import model.subcards.*;
+
 import java.util.ArrayList;
 
-import subcards.Card;
 
 public class SubDeck {
 	public ArrayList<Card> subdeck = new ArrayList<Card>();
