+++ cucumber-test/src/main/java/model/subcards/Card.java
@@ -1,4 +1,4 @@
-package subcards;
+package model.subcards;
 
 import model.tile.Robot;
