+++ cucumber-test/src/main/java/utilities/Player.java
@@ -1,4 +1,4 @@
-package model;
+package utilities;
 
 import java.util.ArrayList;
