+++ cucumber-test/src/test/java/stepsDefinitionGiovanni/GameMasterSteps.java
@@ -5,7 +5,7 @@ import static org.junit.Assert.*;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
-import model.GameMaster;
+import utilities.GameMaster;
 
 public class GameMasterSteps {
